add more columns

+++ database/migrations/2024_04_22_064141_add_column_name_to_table_name.php
@@ -22,7 +22,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('employees', function (Blueprint $table) {
-            $table->dropColumn('age');
+            if (Schema::hasColumn('employees', 'age')) {
+                $table->dropColumn('age');
+            }
         });
     }
 };
